milestone 3 last exercise done

+++ main.c
@@ -12,12 +12,17 @@
 #include <string.h>
 #include <sys/wait.h>
 #include <signal.h>
+#include <sys/epoll.h>
 
 #include "errorHandling.h"
 #include "performConnection.h"
 #include "paramConfig.h"
 #include "shmConnectorThinker.h"
 #include "thinking.h"
+
+
+#define MAX_EVENTS 4
+#define READ_SIZE 80
 
 /*
 #define GAMEKINDNAME "NMMorris"
@@ -27,6 +32,8 @@
 
 void *shmPtr_thinker;
 int *initial_shm_ptr;
+// Thinker -> Connector Pipe
+int tc_pipe[2];
     
 void printAnweisung(){
     printf("-g <GAME-ID> 13-stellige Game-ID\n");
@@ -94,6 +101,15 @@ int getSocketDescriptorAndConnect(PARAM_CONFIG_T* cfg){
 void signalHandler(int signal){
     printf("signalHandler %d\n",signal);
     if(signal == SIGUSR1){
+        //let the thinker think and send it to the connector
+        //TODO: think based on the last servermsg
+        char buff[] = "THINKING\n";
+        // thinker writes to pipe
+        if(write(tc_pipe[1], buff, strlen(buff) + 1) == -1){
+            perror("thinker can't write.\n");
+            exit(0);
+        }
+
         //for test purpose
         printf("thinking... shm_id %d\n",*initial_shm_ptr);
         //Attaching actual shared memory segment with id *initial_shm_ptr for internal communication with Connector
@@ -116,6 +132,10 @@ void signalHandler(int signal){
 
 
 
+/// @brief 
+/// @param argc 
+/// @param argv 
+/// @return 
 int main(int argc,char**argv){
 
     if (argc <= 1)
@@ -193,6 +213,13 @@ int main(int argc,char**argv){
 
     initial_shm_ptr = (int *) attachShm(initial_shm_id);
 
+
+    // creating pipe and error handling
+    if(pipe(tc_pipe) != 0){
+        perror("pipe() failed.\n");
+        return -1;
+    }
+
     //Connector Process
     if((pid = fork()) == 0){
         //Preparing connection to server "sysprak.priv.lab.nm.ifi.lmu.de"
@@ -252,9 +279,87 @@ int main(int argc,char**argv){
             printf("Connector shm_allPlayerInfo[1]>playerNumber: %d\n", shm_allPlayerInfo[1]->playerNumber);
              printf("Connector shm_allPlayerInfo[1]->playerName: %s\n", shm_allPlayerInfo[1]->playerName);
         }   
-        //TEST send signal to thinker
-        kill(gameInfo->idThinker, SIGUSR1);
-        printf("Connector: SIGUSR1 sent\n");
+
+
+        char buffer[READ_SIZE +1];
+        int event_count = 0;
+        //size_t bytes_read;
+        struct epoll_event event_sock, event_pipe, events[MAX_EVENTS];
+
+        int running = 1, epoll_fd = epoll_create1(0);
+
+        //connector closes write
+        close(tc_pipe[1]);
+
+        if(epoll_fd == -1){
+            perror("Failed to create epoll file descriptor.\n");
+            return -1;
+        }
+
+        //adds  pipe file descriptors to epoll to check for changes
+        event_pipe.events = EPOLLIN;
+        event_pipe.data.fd = tc_pipe[0];
+        if(epoll_ctl(epoll_fd, EPOLL_CTL_ADD, tc_pipe[0], &event_pipe)){
+            perror("Failed to add file descriptor to epoll.\n");
+            close(epoll_fd);
+            return -1;
+        }
+
+        //adds socket file descriptor to epoll to check for changes
+        event_sock.events = EPOLLIN;
+        event_sock.data.fd = socketfd;
+        if(epoll_ctl(epoll_fd, EPOLL_CTL_ADD, socketfd, &event_sock)){
+            perror("Failed to add file descriptor to epoll.\n");
+            close(epoll_fd);
+            return -1;
+        }
+        
+
+
+        printf("\nPolling for input...\n");
+        while(running){
+            event_count = epoll_wait(epoll_fd, events, MAX_EVENTS, -1);
+            //Debug ready events
+            //printf("%d ready events\n", event_count);
+
+            for(int i = 0; i < event_count; ++i){
+                //incoming event is from server
+                if(events[i].data.fd == socketfd){
+                    
+                    getServermsg(socketfd);
+                    //Debug Socket 
+                    printf("Read socket : %s\n", serverMsg);
+
+                    //TODO: Send signal to thinker when the server sends a msg
+                    //TEST send signal to thinker
+                    kill(gameInfo->idThinker, SIGUSR1);
+                    printf("Connector: SIGUSR1 sent\n");
+
+                    //Terminate on negative msg
+                    if(strncmp(serverMsg, "-", 1) == 0){
+                        running = 0;
+                    }
+
+                }
+                //incoming event is from pipe
+                else if(events[i].data.fd == tc_pipe[0]){
+                    if(read(tc_pipe[0], &buffer, READ_SIZE + 1) == -1){
+                        perror("connector process can't read from pipe.\n");
+                        return -1;
+                    }
+
+                    sendMsgToServer(socketfd, buffer);  
+                }
+                
+
+            }
+        }
+
+        if(close(epoll_fd)){
+            perror("Failed to close epoll file descriptor.\n");
+            return -1;
+        }
+
 
         //free memory for GAMEINFO *gameInfo
         free(gameInfo);
@@ -265,6 +370,8 @@ int main(int argc,char**argv){
     }
 
     // Thinker Process starts here
+    //close reading pipe for the thinker
+    close(tc_pipe[0]);
     printf("Thinker after fork, shm_id %d\n",*initial_shm_ptr);
     //Attaching actual shared memory segment with id *initial_shm_ptr for internal communication with Connector
     //void *shmPtr_thinker; -> global variable
@@ -295,6 +402,7 @@ int main(int argc,char**argv){
         printf("Thinker shm_allPlayerInfo[1]>playerNumber: %d\n", shm_allPlayerInfo[1]->playerNumber);
     }   
 #endif
+
 
     // avoids orphan and zombie process, wait for child to die
     while(wait(NULL) > 0){
